修改图片压缩方法，修正抠图的图片变形问题

+++ ImageUI/src/main/java/cropper/util/ImageViewUtil.java
@@ -190,8 +190,8 @@ public class ImageViewUtil {
     public static int calculateInSampleSize(int width, int height, int reqWidth, int reqHeight) {
         int inSampleSize = 1;
         if (height > reqHeight || width > reqWidth) {
-            final int halfHeight = height / 2;
-            final int halfWidth = width / 2;
+            final int halfHeight = (int)(height/2) ;
+            final int halfWidth = (int)(width/2);
             while ((halfHeight / inSampleSize) > reqHeight && (halfWidth / inSampleSize) > reqWidth) {
                 inSampleSize *= 2;
             }
